requirements.txt
inflect==6.0.4
praw==7.7.0
python-telegram-bot==20.3
pytz==2023.3
requests==2.31.0
